Ensure that the average word count is an integer not a decimal.

+++ src/com/quollwriter/ui/charts/SessionWordCountChart.java
@@ -461,7 +461,7 @@ public class SessionWordCountChart extends AbstractQuollChart<AbstractViewer>
                 
                 t = String.format (", %s%s target",
                                    (diffAvg < 0 ? "" : "+"),
-                                   Environment.formatNumber (diffAvg));
+                                   Environment.formatNumber ((long) diffAvg));
                 
             }
